requirements.txt
h5py>=2.8.0
Keras>=2.2.2
Keras-Applications>=1.0.4
Keras-Preprocessing>=1.0.2
kiwisolver>=1.0.1
matplotlib>=2.2.3
networkx>=2.2
numpy>=1.15.1
Pillow>=5.2.0
protobuf>=3.6.1
pyparsing>=2.2.0
python-dateutil>=2.7.3
pytz>=2018.5
scikit-image>=0.14.1
scipy>=1.1.0
six>=1.11.0
tensorflow-gpu>=1.10.0
termcolor>=1.1.0
toolz>=0.9.0
